Fix example handler to receive a message
There appears to be an error in the example code for sending and receiving client side messages in Javascript. As written, every message logged to the console is null. In order to make the example work, I needed to add another parameter before the 'message' parameter so that it matches the source code for the onmessage function in vertx-eventbus.js on lines 131 and 141. The onmessage function has two calls which pass two arguments to a handler, and the arguments are '(null, json)'.

Therefore, when my client code matched the example and only had one argument, it was always arriving as null. My proposed fix to the example code allows the proper json message to be received by my client's handler.